repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/13/LLMatic/test_mock_db.py
import unittest
import mock_db

class TestMockDB(unittest.TestCase):
	def setUp(self):
		self.db = mock_db.MockDB()

	def test_online_status(self):
		self.assertTrue(self.db.online)
		self.db.set_online(False)
		self.assertFalse(self.db.online)
		self.db.set_online(True)
		self.assertTrue(self.db.online)

	def test_message_queue(self):
		self.assertEqual(self.db.message_queue, [])
		self.db.queue_message({'message_id': '1', 'text': 'Hello'})
		self.assertEqual(self.db.message_queue, [{'message_id': '1', 'text': 'Hello'}])
		self.db.set_online(True)
		self.assertEqual(self.db.message_queue, [])
		self.assertEqual(self.db.retrieve('1'), {'text': 'Hello'})

# ... rest of the code remains the same ...
